feat(pci-kubernetes): update plan help drawer (#22137)
ref: #TAPC-5552
