+++ packages/manager/apps/pci-kubernetes/public/translations/add/Messages_en_GB.json
@@ -121,9 +121,6 @@
   "kube_add_plan_helper_difference_nodes": "Number of nodes: up to 100 (Free) vs 500 (Standard)",
   "kube_add_plan_helper_difference_az": "Availability zones: 1-AZ (Free) vs 3-AZ (Standard)",
   "kube_add_plan_helper_v2_title": "Which plan to choose for your Kubernetes cluster?",
-  "kube_add_plan_helper_v2_intro_1": "The choice of plan determines the availability of your",
-  "kube_add_plan_helper_v2_intro_control_plane": "Control Plane",
-  "kube_add_plan_helper_v2_intro_2": "Note: once the cluster is created, it is impossible to switch from the Free plan to the Standard plan (and vice versa).",
   "kube_add_plan_helper_v2_free_title": "Free Plan (Gratuit)",
   "kube_add_plan_helper_v2_free_ideal_label": "Ideal for:",
   "kube_add_plan_helper_v2_free_ideal_text": "Testing, development, and learning.",
@@ -136,12 +133,12 @@
   "kube_add_plan_helper_v2_standard_availability_1az": "Single zone (1-AZ): High performance in a single geographical zone (SLA 99.9%)",
   "kube_add_plan_helper_v2_standard_availability_3az": "Multi-zone (3-AZ): Maximum resilience. The cluster remains operational even in the event of a loss of an availability zone (SLA 99.99%)",
   "kube_add_plan_helper_v2_warning_title": "⚠️ Crucial reminder:",
-  "kube_add_plan_helper_v2_warning_text": "If you anticipate that your project will go into production, choose the",
-  "kube_add_plan_helper_v2_warning_plan_standard": "Standard Plan",
-  "kube_add_plan_helper_v2_warning_text_suffix": "now.",
   "kube_add_plan_helper_v2_warning_consequence": "If you choose the Free plan, you will need to completely recreate a new cluster and manually migrate your data to benefit from the Standard later.",
   "kube_add_plan_content_not_available_AZ": "Unavailable for {{mode}}",
   "kube_add_plan_content_not_available_region": "Unavailable for this region",
   "kube_add_cart_total_price": "Total",
-  "kube_add_cart_monthly_price": "Monthly estimate"
+  "kube_add_cart_monthly_price": "Monthly estimate",
+  "kube_add_plan_helper_v2_intro_1": "The choice of plan determines the availability of your <b>Control Plane</b>.",
+  "kube_add_plan_helper_v2_intro_2": "<b>Warning: once the cluster is created, it is impossible to switch from the Free plan to the Standard plan (and vice versa).</b>",
+  "kube_add_plan_helper_v2_warning_text": "If you expect your project to go into production, choose the <b>Standard Plan</b> now."
 }
