+++ packages/manager/apps/pci-kubernetes/public/translations/add/Messages_es_ES.json
@@ -91,7 +91,7 @@
   "kube_add_plan_content_standard_1AZ_SLA": "SLA DEL 99,9 %",
   "kubernetes_add_create_cluster": "Crear mi clúster",
   "kubernetes_add_cluster_name_input_helper": {
-    "introduction": "Debe:",
+    "introduction": "Debe :",
     "char_number": "contener entre 2 y 64 caracteres",
     "char_whitelist": "contener únicamente letras, números, guiones o guiones bajos",
     "first_char": "comenzar con una letra",
@@ -121,9 +121,9 @@
   "kube_add_plan_helper_difference_nodes": "Número de nodos: hasta 100 (Gratis) vs 500 (Estándar)",
   "kube_add_plan_helper_difference_az": "Zonas de disponibilidad: 1-AZ (Gratis) vs 3-AZ (Estándar)",
   "kube_add_plan_helper_v2_title": "¿Qué plan elegir para su clúster de Kubernetes?",
-  "kube_add_plan_helper_v2_intro_1": "La elección del plan determina la disponibilidad de su",
+  "kube_add_plan_helper_v2_intro_1": "La elección del plan determina la disponibilidad de su <b>Control Plane</b>.",
   "kube_add_plan_helper_v2_intro_control_plane": "Plano de Control",
-  "kube_add_plan_helper_v2_intro_2": "Atención: una vez creado el clúster, no es posible cambiar del plan Gratis al plan Estándar (y viceversa).",
+  "kube_add_plan_helper_v2_intro_2": "<b>Atención: una vez creado el clúster, no es posible cambiar del plan Free al plan Standard (y viceversa).</b>",
   "kube_add_plan_helper_v2_free_title": "Plan Gratis (Free)",
   "kube_add_plan_helper_v2_free_ideal_label": "Ideal para:",
   "kube_add_plan_helper_v2_free_ideal_text": "Pruebas, desarrollo y aprendizaje.",
@@ -136,12 +136,10 @@
   "kube_add_plan_helper_v2_standard_availability_1az": "Monozone (1-AZ) : Alto rendimiento en una sola zona geográfica (SLA 99.9%)",
   "kube_add_plan_helper_v2_standard_availability_3az": "Multizona (3-AZ) : Resiliencia máxima. El clúster sigue operativo incluso en caso de pérdida de una zona de disponibilidad (SLA 99.99%)",
   "kube_add_plan_helper_v2_warning_title": "⚠️ Recordatorio crucial :",
-  "kube_add_plan_helper_v2_warning_text": "Si prevé que su proyecto pasará a producción, elija el",
-  "kube_add_plan_helper_v2_warning_plan_standard": "Plan Estándar",
-  "kube_add_plan_helper_v2_warning_text_suffix": "desde ahora.",
   "kube_add_plan_helper_v2_warning_consequence": "Si elige el plan Free, deberá recrear completamente un nuevo clúster y migrar sus datos manualmente para beneficiarse de las ventajas del Standard más tarde.",
   "kube_add_plan_content_not_available_AZ": "No disponible para {{mode}}",
   "kube_add_plan_content_not_available_region": "No disponible para esta región",
   "kube_add_cart_total_price": "Total",
-  "kube_add_cart_monthly_price": "Estimación mensual"
+  "kube_add_cart_monthly_price": "Estimación mensual",
+  "kube_add_plan_helper_v2_warning_text": "Si prevé que su proyecto pasará a producción, elija el <b>Plan Standard</b> ahora mismo."
 }
