+++ packages/manager/apps/pci-kubernetes/public/translations/add/Messages_fr_FR.json
@@ -119,9 +119,8 @@
   "kube_add_plan_helper_difference_nodes": "Nombre de nœuds : jusqu'à 100 (Free) vs 500 (Standard)",
   "kube_add_plan_helper_difference_az": "Zones de disponibilité : 1-AZ (Free) vs 3-AZ (Standard)",
   "kube_add_plan_helper_v2_title": "Quel plan choisir pour votre cluster Kubernetes ?",
-  "kube_add_plan_helper_v2_intro_1": "Le choix du plan détermine la disponibilité de votre",
-  "kube_add_plan_helper_v2_intro_control_plane": "Control Plane",
-  "kube_add_plan_helper_v2_intro_2": "Attention : une fois le cluster créé, il est impossible de passer du plan Free au plan Standard (et inversement).",
+  "kube_add_plan_helper_v2_intro_1": "Le choix du plan détermine la disponibilité de votre <b>Control Plane</b>.",
+  "kube_add_plan_helper_v2_intro_2": "<b>Attention : une fois le cluster créé, il est impossible de passer du plan Free au plan Standard (et inversement).</b>",
   "kube_add_plan_helper_v2_free_title": "Plan Free (Gratuit)",
   "kube_add_plan_helper_v2_free_ideal_label": "Idéal pour :",
   "kube_add_plan_helper_v2_free_ideal_text": "Tests, développement et apprentissage.",
@@ -134,9 +133,7 @@
   "kube_add_plan_helper_v2_standard_availability_1az": "Monozone (1-AZ) : Haute performance dans une seule zone géographique (SLA 99.9%)",
   "kube_add_plan_helper_v2_standard_availability_3az": "Multizone (3-AZ) : Résilience maximale. Le cluster reste opérationnel même en cas de perte d'une zone de disponibilité (SLA 99.99%)",
   "kube_add_plan_helper_v2_warning_title": "⚠️ Rappel crucial :",
-  "kube_add_plan_helper_v2_warning_text": "Si vous prévoyez que votre projet passera en production, choisissez le",
-  "kube_add_plan_helper_v2_warning_plan_standard": "Plan Standard",
-  "kube_add_plan_helper_v2_warning_text_suffix": "dès maintenant.",
+  "kube_add_plan_helper_v2_warning_text": "Si vous prévoyez que votre projet passera en production, choisissez le <b>Plan Standard</b> dès maintenant.",
   "kube_add_plan_helper_v2_warning_consequence": "Si vous choisissez le plan Free, vous devrez recréer entièrement un nouveau cluster et migrer vos données manuellement pour bénéficier des avantages du Standard plus tard.",
   "kube_add_plan_content_not_available_AZ":  "Indisponible pour {{mode}}",
   "kube_add_plan_content_not_available_region":  "Indisponible pour cette région",
