+++ packages/manager/apps/pci-kubernetes/public/translations/add/Messages_it_IT.json
@@ -91,7 +91,7 @@
   "kube_add_plan_content_standard_1AZ_SLA": "99,9% SLA",
   "kubernetes_add_create_cluster": "Creare il mio cluster",
   "kubernetes_add_cluster_name_input_helper": {
-    "introduction": "Deve:",
+    "introduction": "Deve :",
     "char_number": "contenere tra 2 e 64 caratteri",
     "char_whitelist": "contenere solo lettere, numeri, trattini o underscore",
     "first_char": "iniziare con una lettera",
@@ -121,9 +121,6 @@
   "kube_add_plan_helper_difference_nodes": "Numero di nodi: fino a 100 (Gratuito) vs 500 (Standard)",
   "kube_add_plan_helper_difference_az": "Zone di disponibilità: 1-AZ (Gratuito) vs 3-AZ (Standard)",
   "kube_add_plan_helper_v2_title": "Quale piano scegliere per il tuo cluster Kubernetes?",
-  "kube_add_plan_helper_v2_intro_1": "La scelta del piano determina la disponibilità del tuo",
-  "kube_add_plan_helper_v2_intro_control_plane": "Control Plane",
-  "kube_add_plan_helper_v2_intro_2": "Attenzione: una volta creato il cluster, non è possibile passare dal piano Gratuito al piano Standard (e viceversa).",
   "kube_add_plan_helper_v2_free_title": "Piano Gratuito",
   "kube_add_plan_helper_v2_free_ideal_label": "Ideale per",
   "kube_add_plan_helper_v2_free_ideal_text": "Test, sviluppo e apprendimento.",
@@ -136,12 +133,12 @@
   "kube_add_plan_helper_v2_standard_availability_1az": "Monozona (1-AZ) : Alta prestazione in un'unica area geografica (SLA 99.9%)",
   "kube_add_plan_helper_v2_standard_availability_3az": "Multizona (3-AZ) : Massima resilienza. Il cluster rimane operativo anche in caso di perdita di un'area di disponibilità (SLA 99.99%)",
   "kube_add_plan_helper_v2_warning_title": "⚠️ Promemoria cruciale :",
-  "kube_add_plan_helper_v2_warning_text": "Se prevedi che il tuo progetto passerà in produzione, scegli il",
-  "kube_add_plan_helper_v2_warning_plan_standard": "Piano Standard",
-  "kube_add_plan_helper_v2_warning_text_suffix": "da ora.",
   "kube_add_plan_helper_v2_warning_consequence": "Se scegli il piano Free, dovrai ricreare completamente un nuovo cluster e migrare manualmente i tuoi dati per beneficiare dei vantaggi dello Standard in seguito.",
   "kube_add_plan_content_not_available_AZ": "Non disponibile per {{mode}}",
   "kube_add_plan_content_not_available_region": "Non disponibile per questa regione",
   "kube_add_cart_total_price": "Totale",
-  "kube_add_cart_monthly_price": "Stima mensile"
+  "kube_add_cart_monthly_price": "Stima mensile",
+  "kube_add_plan_helper_v2_intro_1": "La scelta del piano determina la disponibilità del <b>Control Plane</b>.",
+  "kube_add_plan_helper_v2_intro_2": "<b>Attenzione : una volta creato il cluster, non è possibile passare dal piano Free al piano Standard (e viceversa).</b>",
+  "kube_add_plan_helper_v2_warning_text": "Se prevedi che il tuo progetto passerà in produzione, scegli il <b>Piano Standard</b> adesso."
 }
