+++ packages/manager/apps/pci-kubernetes/public/translations/add/Messages_pl_PL.json
@@ -91,7 +91,7 @@
   "kube_add_plan_content_standard_1AZ_SLA": "SLA 99,9%",
   "kubernetes_add_create_cluster": "Utwórz mój klaster",
   "kubernetes_add_cluster_name_input_helper": {
-    "introduction": "Musisz:",
+    "introduction": "Musisz :",
     "char_number": "zawierać od 2 do 64 znaków",
     "char_whitelist": "zawierać tylko litery, cyfry, myślniki lub podkreślenia",
     "first_char": "zaczynać się od litery",
@@ -121,9 +121,6 @@
   "kube_add_plan_helper_difference_nodes": "Liczba węzłów: do 100 (Darmowy) vs 500 (Standard)",
   "kube_add_plan_helper_difference_az": "Strefy dostępności: 1-AZ (Darmowy) vs 3-AZ (Standard)",
   "kube_add_plan_helper_v2_title": "Jaki plan wybrać dla swojego klastra Kubernetes?",
-  "kube_add_plan_helper_v2_intro_1": "Wybór planu określa dostępność Twojego",
-  "kube_add_plan_helper_v2_intro_control_plane": "Control Plane",
-  "kube_add_plan_helper_v2_intro_2": "Uwaga: po utworzeniu klastra nie można przejść z planu Darmowego na plan Standard (i odwrotnie).",
   "kube_add_plan_helper_v2_free_title": "Plan Darmowy (Gratuit)",
   "kube_add_plan_helper_v2_free_ideal_label": "Idealny dla:",
   "kube_add_plan_helper_v2_free_ideal_text": "Testy, rozwój i nauka.",
@@ -136,12 +133,12 @@
   "kube_add_plan_helper_v2_standard_availability_1az": "Monozon (1-AZ) : Wysoka wydajność w jednej strefie geograficznej (SLA 99.9%)",
   "kube_add_plan_helper_v2_standard_availability_3az": "Multizone (3-AZ) : Maksymalna odporność. Klaster pozostaje operacyjny nawet w przypadku utraty jednej strefy dostępności (SLA 99.99%)",
   "kube_add_plan_helper_v2_warning_title": "⚠️ Ważne przypomnienie :",
-  "kube_add_plan_helper_v2_warning_text": "Jeśli planujesz, że Twój projekt przejdzie do produkcji, wybierz",
-  "kube_add_plan_helper_v2_warning_plan_standard": "Plan Standard",
-  "kube_add_plan_helper_v2_warning_text_suffix": "już teraz.",
   "kube_add_plan_helper_v2_warning_consequence": "Jeśli wybierzesz plan Free, będziesz musiał całkowicie odtworzyć nowy klaster i ręcznie przenieść swoje dane, aby skorzystać z zalet Standard później.",
   "kube_add_plan_content_not_available_AZ": "Niedostępne dla {{mode}}",
   "kube_add_plan_content_not_available_region": "Niedostępne dla tego regionu",
   "kube_add_cart_total_price": "Razem",
-  "kube_add_cart_monthly_price": "Szacowanie na miesiąc"
+  "kube_add_cart_monthly_price": "Szacowanie na miesiąc",
+  "kube_add_plan_helper_v2_intro_1": "Wybór planu określa dostępność twojego <b>Control Plane</b>.",
+  "kube_add_plan_helper_v2_intro_2": "<b>Uwaga : po utworzeniu klastra nie można przejść z planu Free na plan Standard (i odwrotnie).</b>",
+  "kube_add_plan_helper_v2_warning_text": "Jeśli planujesz, że twój projekt przejdzie do produkcji, wybierz <b>Plan Standard</b> już teraz."
 }
