+++ packages/manager/apps/pci-kubernetes/public/translations/add/Messages_pt_PT.json
@@ -91,11 +91,11 @@
   "kube_add_plan_content_standard_1AZ_SLA": "99,9% SLA",
   "kubernetes_add_create_cluster": "Criar meu cluster",
   "kubernetes_add_cluster_name_input_helper": {
-    "introduction": "Deve:",
+    "introduction": "Deve :",
     "char_number": "conter entre 2 e 64 caracteres",
-    "char_whitelist": "conter apenas letras, números, hífens ou underscores",
-    "first_char": "começar com uma letra",
-    "last_char": "terminar com um caractere alfanumérico (a-z, A-Z, 0-9)"
+    "char_whitelist": "conter apenas letras, números, hífens ou sublinhados",
+    "first_char": "começar por uma letra",
+    "last_char": "terminar com um carácter alfanumérico (a-z, A-Z, 0-9)"
   },
   "kubernetes_add_cluster_name_input_error": "O nome do cluster é inválido",
   "kubernetes_add_cluster_name_input_error_empty": "O nome do cluster é obrigatório",
@@ -121,9 +121,6 @@
   "kube_add_plan_helper_difference_nodes": "Número de nós: até 100 (Gratuito) vs 500 (Standard)",
   "kube_add_plan_helper_difference_az": "Zonas de disponibilidade: 1-AZ (Gratuito) vs 3-AZ (Standard)",
   "kube_add_plan_helper_v2_title": "Qual plano escolher para o seu cluster Kubernetes?",
-  "kube_add_plan_helper_v2_intro_1": "A escolha do plano determina a disponibilidade do seu",
-  "kube_add_plan_helper_v2_intro_control_plane": "Control Plane",
-  "kube_add_plan_helper_v2_intro_2": "Atenção: uma vez que o cluster é criado, não é possível mudar do plano Gratuito para o plano Standard (e vice-versa).",
   "kube_add_plan_helper_v2_free_title": "Plano Gratuito",
   "kube_add_plan_helper_v2_free_ideal_label": "Ideal para:",
   "kube_add_plan_helper_v2_free_ideal_text": "Testes, desenvolvimento e aprendizado.",
@@ -136,12 +133,12 @@
   "kube_add_plan_helper_v2_standard_availability_1az": "Monozone (1-AZ) : Alto desempenho em uma única zona geográfica (SLA 99.9%)",
   "kube_add_plan_helper_v2_standard_availability_3az": "Multizone (3-AZ) : Resiliência máxima. O cluster permanece operacional mesmo em caso de perda de uma zona de disponibilidade (SLA 99.99%)",
   "kube_add_plan_helper_v2_warning_title": "⚠️ Lembrete crucial :",
-  "kube_add_plan_helper_v2_warning_text": "Se você prevê que seu projeto passará para produção, escolha o",
-  "kube_add_plan_helper_v2_warning_plan_standard": "Plano Standard",
-  "kube_add_plan_helper_v2_warning_text_suffix": "já disponível.",
   "kube_add_plan_helper_v2_warning_consequence": "Se você escolher o plano Free, precisará recriar completamente um novo cluster e migrar seus dados manualmente para aproveitar os benefícios do Standard mais tarde.",
   "kube_add_plan_content_not_available_AZ": "Indisponível para {{mode}}",
   "kube_add_plan_content_not_available_region": "Indisponível para esta região",
   "kube_add_cart_total_price": "Total",
-  "kube_add_cart_monthly_price": "Estimativa mensal"
+  "kube_add_cart_monthly_price": "Estimativa mensal",
+  "kube_add_plan_helper_v2_intro_1": "A escolha do plano determina a disponibilidade do seu <b>Control Plane</b>.",
+  "kube_add_plan_helper_v2_intro_2": "<b>Atenção : uma vez que o cluster é criado, não é possível mudar do plano Free para o plano Standard (e vice-versa).</b>",
+  "kube_add_plan_helper_v2_warning_text": "Se você prevê que o seu projeto passará para produção, escolha o <b>Plano Standard</b> agora mesmo."
 }
